rename to LoginDetectionJavascriptInterface
