+++ app/src/androidTest/java/com/duckduckgo/app/browser/logindetection/JSLoginDetectorTest.kt
@@ -45,7 +45,7 @@ class JSLoginDetectorTest {
     fun whenAddLoginDetectionThenJSInterfaceAdded() = coroutinesTestRule.runBlocking {
         val webView = spy(WebView(InstrumentationRegistry.getInstrumentation().targetContext))
         testee.addLoginDetection(webView) {}
-        verify(webView).addJavascriptInterface(any<LoginDetectionJsInterface>(), eq(LOGIN_DETECTION_JS_INTERFACE_NAME))
+        verify(webView).addJavascriptInterface(any<LoginDetectionJavascriptInterface>(), eq(LOGIN_DETECTION_JS_INTERFACE_NAME))
     }
 
     @UiThreadTest
